Fixed: Download notification was disappearing when the application is in background.
