+++ core/src/main/AndroidManifest.xml
@@ -90,5 +90,8 @@
       android:name=".error.DiagnosticReportActivity"
       android:exported="false" />
     <service android:name=".read_aloud.ReadAloudService" />
+    <service
+      android:name=".downloader.downloadManager.DownloadMonitorService"
+      android:foregroundServiceType="dataSync" />
   </application>
 </manifest>
