+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/components/CoreComponent.kt
@@ -52,7 +52,6 @@ import org.kiwix.kiwixmobile.core.di.modules.NetworkModule
 import org.kiwix.kiwixmobile.core.di.modules.SearchModule
 import org.kiwix.kiwixmobile.core.downloader.Downloader
 import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadManagerBroadcastReceiver
-import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadNotificationActionsBroadcastReceiver
 import org.kiwix.kiwixmobile.core.error.ErrorActivity
 import org.kiwix.kiwixmobile.core.main.KiwixWebView
 import org.kiwix.kiwixmobile.core.reader.ZimFileReader
@@ -117,8 +116,6 @@ interface CoreComponent {
   fun mutex(): Mutex
 
   fun downloadManagerBroadCastReceiver(): DownloadManagerBroadcastReceiver
-  fun downloadNotificationActionBroadCastReceiver(): DownloadNotificationActionsBroadcastReceiver
-
   fun inject(application: CoreApp)
   fun inject(kiwixWebView: KiwixWebView)
   fun inject(storageSelectDialog: StorageSelectDialog)
