+++ core/src/main/java/org/kiwix/kiwixmobile/core/di/components/CoreServiceComponent.kt
@@ -23,12 +23,14 @@ import dagger.BindsInstance
 import dagger.Subcomponent
 import org.kiwix.kiwixmobile.core.di.CoreServiceScope
 import org.kiwix.kiwixmobile.core.di.modules.CoreServiceModule
+import org.kiwix.kiwixmobile.core.downloader.downloadManager.DownloadMonitorService
 import org.kiwix.kiwixmobile.core.read_aloud.ReadAloudService
 
 @Subcomponent(modules = [CoreServiceModule::class])
 @CoreServiceScope
 interface CoreServiceComponent {
   fun inject(readAloudService: ReadAloudService)
+  fun inject(downloadMonitorService: DownloadMonitorService)
 
   @Subcomponent.Builder
   interface Builder {
